Fix iOS build errors, complex view type-check timeouts, and dependency imports
- UserContributionsView: Add missing FirebaseAuth import

- SponsorsView: Fix malformed ZStack/NavigationView hierarchy

- MountainDetailView: Fix UInt to UInt64 type mismatch for GMSPlaceField

- AdminModerationQueueView: Add missing emptyTabMessage view builder

- LocationHelper: Add Equatable extension for CLLocationCoordinate2D

- AddCustomMountainView: Break down complex Form into ViewBuilder sections to fix SwiftUI compiler timeouts; correctly use MountainsViewModel.submitCustomMountain for new peaks

+++ PataGilid/Views/MountainDetailView.swift
@@ -1107,7 +1107,7 @@ struct InteractiveCoordinatePickerView: View {
                 return
             }
             
-            let fields: GMSPlaceField = GMSPlaceField(rawValue: UInt(GMSPlaceField.coordinate.rawValue))
+            let fields: GMSPlaceField = GMSPlaceField(rawValue: UInt64(GMSPlaceField.coordinate.rawValue))
             GMSPlacesClient.shared().fetchPlace(fromPlaceID: firstResult.placeID, placeFields: fields, sessionToken: nil) { (place, error) in
                 isSearching = false
                 if let error = error {
